requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
jinja2==3.1.2
python-dotenv==1.0.0
httpx==0.25.1
python-multipart==0.0.6
openai==1.2.4
ansible-lint==6.22.0
pydantic==2.4.2
pytest==7.4.3
bootstrap-flask==2.3.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
